feat(product): 添加商品管理功能- 新增商品列表页面，支持查看、添加、编辑和删除商品 - 实现商品相关的 API接口，包括获取商品列表、添加商品、删除商品和更新商品 - 新增商品日志记录功能，记录商品操作 -优化商品相关的数据结构和接口定义

+++ backend/src/com/ddl/entity/ProductLog.java
@@ -10,9 +10,9 @@ import java.time.LocalDate;
 @Table("product_log")
 public class ProductLog {
     @Id
-    private long id;                                //日志ID
+    private int id;                                //日志ID
 
-    private long itemId;                            //商品ID
+    private int itemId;                            //商品ID
     private String itemName;                        //商品名称
     private String action;                          //修改操作
     private LocalDate createdAt;                    //创建时间
